JBIDE-22305 fix verifyTarget to provide more debugging when a dep is missind; add changeTargetURLs wrapper that can be called from Jenkins to push new URLs from upstream into .target files and run validation on them; add p2diff-check.sh

+++ util/verifyTarget.sh
@@ -70,6 +70,7 @@ MVN="mvn"
 includeSources="-Dmirror-target-to-repo.includeSources=true" # by default, include sources
 INSTALLSCRIPT=/tmp/installFromTarget.sh
 LOG_GREP_INCLUDES="BUILD FAILURE|Only one of the following|Missing requirement|Unresolved requirement|IllegalArgumentException|Could not resolve|could not be found|being installed|Cannot satisfy dependency|FAILED"
+LOG_GREP_INCLUDES2="TargetDefinitionResolutionException|Could not find"
 LOG_GREP_EXCLUDES="Failed to execute goal org.jboss.tools.tycho-plugins:target-platform-utils|Checksum validation failed, no checksums available from the repository"
 #BASEDIR=`pwd`
 #ECLIPSEZIP=${HOME}/tmp/Eclipse_Bundles/eclipse-jee-neon-M2-linux-gtk-x86_64.tar.gz
@@ -163,12 +164,14 @@ for PROJECT in $PROJECTS; do echo "Process $PROJECT ..."
   echo ""
   pushd ${WORKDIR}
 
-  for tf in *.target; do 
-    logfile=/tmp/fix-versions_${tf}_log_${PROJECT}_${NOW}.txt
-    echo "${MVN} -U org.jboss.tools.tycho-plugins:target-platform-utils:${targetplatformutilsversion}:fix-versions -DtargetFile=${tf}" | tee $logfile
-    ${MVN} -U -c org.jboss.tools.tycho-plugins:target-platform-utils:${targetplatformutilsversion}:fix-versions -DtargetFile=${tf} | tee -a $logfile
-    egrep -i -v "$LOG_GREP_EXCLUDES" $logfile | egrep -i -A2 "$LOG_GREP_INCLUDES"; if [[ "$?" == "0" ]]; then break 2; fi
-    if [[ -f ${tf}_fixedVersion.target ]]; then rm -f ${tf} *_update_hints.txt; mv -f ${tf}{_fixedVersion.target,}; fi
+  for tf in *.target; do
+    if [[ ${tf/_fixedVersion.target} == ${tf} ]]; then 
+      logfile=/tmp/fix-versions_${tf}_log_${PROJECT}_${NOW}.txt
+      echo "${MVN} -U org.jboss.tools.tycho-plugins:target-platform-utils:${targetplatformutilsversion}:fix-versions -DtargetFile=${tf}" | tee $logfile
+      ${MVN} -U -c org.jboss.tools.tycho-plugins:target-platform-utils:${targetplatformutilsversion}:fix-versions -DtargetFile=${tf} | tee -a $logfile
+      egrep -i -v "$LOG_GREP_EXCLUDES" $logfile | egrep -i -A2 "$LOG_GREP_INCLUDES"; if [[ "$?" == "0" ]]; then break 2; fi
+      if [[ -f ${tf}_fixedVersion.target ]]; then rm -f ${tf} *_update_hints.txt; mv -f ${tf}{_fixedVersion.target,}; fi
+    fi
   done
   popd
 
@@ -184,9 +187,9 @@ for PROJECT in $PROJECTS; do echo "Process $PROJECT ..."
   # TODO: if you removed IUs, be sure to do a `mvn clean install`, rather than just a `mvn install`; process will be much longer but will guarantee metadata is correct
   pushd ${WORKSPACE}
   logfile=/tmp/resolve_log_${PROJECT}_${NOW}.txt
-  echo "${MVN} install -P${PROFILE} -DtargetRepositoryUrl=file://${WORKDIR}/target/${REPODIR}/ ${includeSources}" | tee $logfile
-  ${MVN} install -P${PROFILE} -DtargetRepositoryUrl=file://${WORKDIR}/target/${REPODIR}/ ${includeSources} | tee -a $logfile
-  egrep -i -v "$LOG_GREP_EXCLUDES" $logfile | egrep -i -A2 "$LOG_GREP_INCLUDES"; if [[ "$?" == "0" ]]; then break 2; fi
+  echo "${MVN} install -P${PROFILE} -DtargetRepositoryUrl=file://${WORKDIR}/target/${REPODIR}/ ${includeSources} -X" | tee $logfile
+  ${MVN} install -P${PROFILE} -DtargetRepositoryUrl=file://${WORKDIR}/target/${REPODIR}/ ${includeSources} -X | tee -a $logfile
+  egrep -i -v "$LOG_GREP_EXCLUDES" $logfile | egrep -i -A2 "$LOG_GREP_INCLUDES|$LOG_GREP_INCLUDES2"; if [[ "$?" == "0" ]]; then break 2; fi
 
   popd
